feat(rcs): migrate all RCS APIs from rainyun-go-sdk
- 35 endpoints across 14 files, response structs split for reuse
- Reuse RcsRecord/Node/Plan/OsInfo/UsageData across list/detail/create
- Fix source bugs: {id} path placeholders, missing / in firewall pos path,
  missing body in ChargeRcsTrafic, resp vs &resp
- Add HTTPMethod_PATCH, common.BasicOperationResponse, VncConnectionInfo

+++ apis/rcs/get_rcs_list.go
@@ -47,9 +47,9 @@ type UsageData struct {
 	UsedMem     int           `json:"UsedMem"`
 	Disks       DiskUsageData `json:"Disks"`
 	DiskRead    int           `json:"DiskRead"`
-	DiskWrite   int           `json:"DiskWrite"`
+	DiskWrite   float64       `json:"DiskWrite"`
 	NetOut      float64       `json:"NetOut"`
-	NetIn       int           `json:"NetIn"`
+	NetIn       float64       `json:"NetIn"`
 	SmartHealth any           `json:"SmartHealth"`
 	SmartTemp   int           `json:"SmartTemp"`
 	UpdateTime  int           `json:"UpdateTime"`
